do daily 2594

+++ src/TwoSum.cpp
@@ -4,7 +4,7 @@
 #include "solution.h"
 
 // Problem: TwoSum
-// Description: Given an array of integers nums and an integer target,
+// Description: Given an array of integers nums and an integer target, 
 // return indices of the two numbers such that they add up to target.
 
 class TwoSumSolution : public Solution {
@@ -12,7 +12,7 @@ public:
     std::vector<int> twoSum(std::vector<int>& nums, int target) {
         std::unordered_map<int, int> map;
         
-        for (int i = 0; i < nums.size(); i++) {
+        for (int i = 0; i < static_cast<int>(nums.size()); i++) {
             int complement = target - nums[i];
             if (map.find(complement) != map.end()) {
                 return {map[complement], i};
